Fix link/c.c truncating 16-bit reloc fixups to 8 bits + add C-level test
link/c.c:667 had a stray (uint8_t) cast in the 16-bit relocation
fixup path:

    conv_u16tob((uint8_t)nval, fixloc);

`conv_u16tob` expects a uint16_t, and the other reloc sizes (1, 3, 4)
pass the value through without truncation. This cast meant that every
16-bit reloc landed in the binary with its high byte zeroed, which
broke any symbol whose address didn't fit in 8 bits.

Symptom on hello.com: zcrtcpm.as does

    ld   de,__Lbss
    ld   hl,__Hbss
    sbc  hl,de              ; size of uninitialized data area
    ld   c,l
    ld   b,h
    ld   (hl),0
    ldir                    ; zero out BSS

With __Lbss/__Hbss truncated to 0x00FC / 0x0009, that LDIR clears
~64KB of garbage memory, the program's PC then wanders into the
freshly-zeroed area, and you get the runaway "Unrecognized BDOS-
Function 132" trace we were seeing. Cast fixed -> __Lbss=0x10FC,
__Hbss=0x1309, hello.com runs cleanly under cpm_emu and prints
"hello world".

Adds test/hello.c covering the full zc3 pipeline (cpp_new3 -> p1x3 ->
cgen3 -> optim3 -> zasx3 -> linq3 + libc startup) to keep this
regression caught.

+++ link/c.c
@@ -664,7 +664,7 @@ void fixup(int offset, uint16_t relocType, uint32_t delta) {
     case 2:
         nval  = delta + (long)conv_letoi16(fixloc);
         nmask = 0xffff;
-        conv_u16tob((uint8_t)nval, fixloc);
+        conv_u16tob((uint16_t)nval, fixloc);
         break;
     case 3:
         nval = delta + conv_btou24(fixloc);
